% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_model_matrix}
\alias{as_model_matrix}
\title{Transform a data frame}
\usage{
as_model_matrix(df, intercept = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{intercept}{Whether to have a column with all \code{1}s. Default is \code{FALSE}.}
}
\value{
A matrix.
}
\description{
Transform a data frame into a matrix using one hot encoding.
}
\examples{
mat <- as_model_matrix(iris)
str(mat)
}
