% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigutilsr-package.R
\docType{package}
\name{bigutilsr-package}
\alias{bigutilsr}
\alias{bigutilsr-package}
\title{bigutilsr: Utility Functions for Large-scale Data}
\description{
Utility functions for large-scale data. For now, package 'bigutilsr'
    mainly includes functions for outlier detection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/privefl/bigutilsr}
  \item Report bugs at \url{https://github.com/privefl/bigutilsr/issues}
}

}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}
\keyword{internal}
