% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdpca.R
\name{pca_adjust}
\alias{pca_adjust}
\title{Adjusting shrinkage in PC scores}
\usage{
pca_adjust(train.eval, p, n, test.scores, method = c("d.gsp", "l.gsp",
  "osp"), n.spikes.max, ncores = 1)
}
\arguments{
\item{train.eval}{
Numeric vector containing the sample eigenvalues. The vector must have dimension \code{n} or \code{n-1}, it may be unordered.
}

\item{p}{
The number of features.
}

\item{n}{
The number of training samples.
}

\item{test.scores}{
An \eqn{m\times k} matrix or data frame containing the first \eqn{k} predicted PC scores of \eqn{m} many test samples.
}

\item{method}{
String specifying the estimation method. Possible values are "\code{d.gsp}" (default),"\code{l.gsp}" and "\code{osp}".
}

\item{n.spikes.max}{
Upper bound of the number of distant spikes in the population. Optional, but needed if \code{n.spikes} is not specified. Ignored if \code{n.spikes} is specified.
}

\item{ncores}{Number of cores to be used. You can e.g. use \code{\link[=nb_cores]{nb_cores()}}.}
}
\value{
A matrix containing the bias-adjusted PC scores.
The dimension of the matrix is the same as the dimension of \code{test.scores}.

Also, an attribute \code{attr(*, "shrinkage")} containing the shrinkage factors.
Note that the number of shrinkage factors can be smaller than the number of
columns of \code{test.scores}; it corresponds to the estimated number of spikes.
}
\description{
Adjusts the shrinkage bias in the predicted PC scores based on the estimated shrinkage factors.
}
