% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdpca.R
\name{pca_nspike}
\alias{pca_nspike}
\title{Number of spikes in PCA}
\usage{
pca_nspike(eigval, nb_in_bin = 10)
}
\arguments{
\item{eigval}{Eigenvalues (squared singular values).}

\item{nb_in_bin}{Mean number of values in each bin of the histogram.}
}
\value{
The estimated number of distant spikes.
}
\description{
Estimate the number of distant spikes based on the histogram of eigenvalues.
}
\seealso{
\code{\link[hdpca:select.nspike]{hdpca::select.nspike()}}
}
