% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.R
\name{predict.procrustes}
\alias{predict.procrustes}
\title{Predict method}
\usage{
\method{predict}{procrustes}(object, X, ...)
}
\arguments{
\item{object}{Object of class \code{procrustes}.}

\item{X}{New matrix to transform.}

\item{...}{Not used.}
}
\value{
\code{X}, transformed.
}
\description{
Predict method for class \code{procrustes}.
}
\seealso{
\code{\link[=procrustes]{procrustes()}}.
}
