% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollmean.R
\name{rollmean}
\alias{rollmean}
\title{Gaussian smoothing}
\usage{
rollmean(x, size)
}
\arguments{
\item{x}{Numeric vector.}

\item{size}{Radius of the smoothing (smaller than half of the length of \code{x}).}
}
\value{
Numeric vector of the same length as \code{x}, smoothed.
}
\description{
Gaussian smoothing
}
\examples{
(x <- rnorm(10))
rollmean(x, 3)
}
