% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{covRob}
\alias{covRob}
\title{Deprecated}
\usage{
covRob(data, ...)
}
\arguments{
\item{data}{A matrix.}

\item{...}{Not used.}
}
\description{
Deprecated
}
\seealso{
\code{\link[bigutilsr:covrob_ogk]{bigutilsr::covrob_ogk()}} \code{\link[bigutilsr:dist_ogk]{bigutilsr::dist_ogk()}}
}
