% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdpca.R
\name{pca_nspike}
\alias{pca_nspike}
\title{Number of spikes in PCA}
\usage{
pca_nspike(eigval, nboot = 1000)
}
\arguments{
\item{eigval}{Eigenvalues (squared singular values).}

\item{nboot}{Number of bootstrap replicates. Default is \code{1000}.}
}
\value{
The estimated number of distant spikes.
}
\description{
Estimate the number of distant spikes based on the histogram of eigenvalues.
}
\seealso{
\code{\link[hdpca:select.nspike]{hdpca::select.nspike()}}
}
