% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdpca.R
\name{pca_nspike}
\alias{pca_nspike}
\title{Number of spikes in PCA}
\usage{
pca_nspike(eigval, breaks = "FD", nboot = 100)
}
\arguments{
\item{eigval}{Eigenvalues (squared singular values).}

\item{breaks}{Same parameter as for \code{hist()}. Default uses a robust version
of Scott's rule. You can use \code{"FD"} or \code{nclass.FD} for a bit more bins.}

\item{nboot}{Number of bootstrap replicates to estimate limits more robustly.
Default is \code{NULL} (no bootstrap, even if \strong{I would recommend to use it}).}
}
\value{
The estimated number of distant spikes.
}
\description{
Estimate the number of distant spikes based on the histogram of eigenvalues.
}
\seealso{
\code{\link[hdpca:select.nspike]{hdpca::select.nspike()}}
}
