% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.R
\name{predict.Procrustes}
\alias{predict.Procrustes}
\title{Predict method}
\usage{
\method{predict}{Procrustes}(object, X, ...)
}
\arguments{
\item{object}{Object of class \code{Procrustes}.}

\item{X}{New matrix to transform.}

\item{...}{Not used.}
}
\value{
\code{X}, transformed.
}
\description{
Predict method for class \code{Procrustes}.
}
\seealso{
\code{\link[=procrustes]{procrustes()}}.
}
